v2 restaurant 구현

+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/main/restaurant/repository/RestaurantV2Repository.kt
@@ -3,6 +3,6 @@ package siksha.wafflestudio.core.domain.main.restaurant.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import siksha.wafflestudio.core.domain.main.restaurant.data.RestaurantV2
 
-interface RestaurantV2Repository : JpaRepository<RestaurantV2, Long> {
+interface RestaurantV2Repository : JpaRepository<RestaurantV2, Int> {
     fun findByName(name: String): RestaurantV2?
 }
